+++ bht/SUTime_processing.py
@@ -1,6 +1,8 @@
 import re
 import json
-from datetime import date, datetime
+from datetime import date, datetime, timedelta
+
+import dateutil.parser as parser
 
 from bht.bht_logging import init_logger
 from tools import RawDumper
@@ -598,7 +600,9 @@ def SUTime_transform(current_OCR_folder):
         except:
             continue
 
-    raw_dumper.dump_to_raw(JSON_list, "Deal type TIME", current_OCR_folder)
+    raw_dumper.dump_to_raw(
+        JSON_list, "Change type TIME to DURATION", current_OCR_folder
+    )
 
     # resolution of all XXXX
     compteur_dicts = 0
@@ -650,9 +654,7 @@ def SUTime_transform(current_OCR_folder):
                 dicts["value"] += "-15"
         compteur_dicts += 1
 
-    raw_dumper.dump_to_raw(
-        JSON_list, "begin -01, end-28, DATE-15, TIME-15", current_OCR_folder
-    )
+    raw_dumper.dump_to_raw(JSON_list, "Date rewriting", current_OCR_folder)
 
     compteur = 0
     for dicts in JSON_list:
@@ -709,6 +711,21 @@ def SUTime_transform(current_OCR_folder):
         compteur += 1
 
     raw_dumper.dump_to_raw(JSON_list, "DURATION gets Nearest date", current_OCR_folder)
+
+    # Change DATE to DURATION
+    for elmnt in JSON_list:
+        if elmnt["type"] == "DATE":
+            date_string = elmnt["timex-value"]
+            try:
+                begin_date = parser.parse(date_string)
+            except parser.ParserError:
+                continue
+            end_date = begin_date + timedelta(days=1) - timedelta(seconds=1)
+            elmnt["value"]={"begin": begin_date.isoformat(), "end": end_date.isoformat()}
+            elmnt["type"] = "DURATION"
+    raw_dumper.dump_to_raw(
+        JSON_list, "Change type DATE to DURATION", current_OCR_folder
+    )
 
     # CLEAR EMPTY
     # At the end of the treatment, deletion of all SUTime results which is not a DURATION
@@ -775,6 +792,8 @@ def SUTime_transform(current_OCR_folder):
         if dicts["type"] != "DURATION":
             dicts.clear()
     JSON_list = [i for i in JSON_list if i != {}]
+
+    raw_dumper.dump_to_raw(JSON_list, "Remove not DURATION", current_OCR_folder)
 
     file = open(
         current_OCR_folder + "/" + "res_sutime_2.json", "w"