+++ tests/conftest.py
@@ -10,20 +10,27 @@ from web import db as _db
 from web.main.routes import pdf_to_db
 from web.models import Paper, HpEvent, BhtFileType
 
+skip_selenium = pytest.mark.skipif(
+    os.environ.get("BHT_SKIP_SELENIUM") is not None
+    and os.environ.get("BHT_SKIP_SELENIUM"),
+    reason="SELENIUM Skipping",
+    )
+
 skip_bht = pytest.mark.skipif(
-    os.environ.get("BHT_DONTSKIPBHT") is None or not os.environ.get("BHT_DONTSKIPBHT"),
+    os.environ.get("BHT_SKIP_BHT") is not None
+    and os.environ.get("BHT_SKIP_BHT"),
     reason="BHT skipping (too long)",
 )
 
 skip_istex = pytest.mark.skipif(
-    os.environ.get("BHT_DONTSKIPISTEX") is None
-    or not os.environ.get("BHT_DONTSKIPISTEX"),
+    os.environ.get("BHT_SKIP_ISTEX") is not None
+    and os.environ.get("BHT_SKIP_ISTEX"),
     reason="ISTEX skipping (no auth)",
 )
 
 skip_slow_test = pytest.mark.skipif(
-    os.environ.get("BHT_SKIPSLOWTESTS") is not None
-    and os.environ.get("BHT_SKIPSLOWTESTS"),
+    os.environ.get("BHT_SKIP_SLOW") is not None
+    and os.environ.get("BHT_SKIP_SLOW"),
     reason="Slow test skipping",
 )
 
@@ -71,7 +78,7 @@ def db(app):
 def paper_with_txt(paper_for_test, txt_for_test):
     """Add a paper's with catalog to db"""
     paper_for_test.set_file_path(txt_for_test, BhtFileType.TXT)
-    paper_for_test.set_doi( "10.1002/jgra.50537")
+    paper_for_test.set_doi("10.1002/jgra.50537")
     _db.session.add(paper_for_test)
     _db.session.commit()
     #
@@ -103,6 +110,24 @@ def paper_for_test(pdf_for_test):
     if paper is not None:
         _db.session.delete(paper)
         _db.session.commit()
+
+
+@pytest.fixture(scope="module")
+def cleaned_for_test():
+    test_cleaned_file_orig = os.path.join(
+        current_app.config["BHT_RESOURCES_DIR"],
+        "CAAEFDA40653763CC8E603A982B2405E1ED646DA.cleaned",
+    )
+    test_cleaned_file_dest = os.path.join(
+        current_app.config["BHT_PAPERS_DIR"],
+        "CAAEFDA40653763CC8E603A982B2405E1ED646DA.cleaned",
+    )
+    shutil.copy(test_cleaned_file_orig, test_cleaned_file_dest)
+    #
+    yield test_cleaned_file_dest
+    #
+    if os.path.isfile(test_cleaned_file_dest):
+        os.remove(test_cleaned_file_dest)
 
 
 @pytest.fixture(scope="module")