+++ tests/unit/test_entities_finder.py
@@ -2,7 +2,6 @@
 # Unit Tests for the big entities_finder file
 #
 import os.path
-from pprint import pprint
 
 from bht.Entities_finder import (
     entities_finder,
@@ -21,7 +20,7 @@ from bht.databank_reader import DataBank, DataBankSheet
 
 class TestEntitiesFinder:
     def test_entities_finder_method(self, ocr_dir_test):
-        entities_finder(ocr_dir_test, DOI="10.1002/2015GL064052")
+        entities_finder(ocr_dir_test, doc_meta_info={"doi": "10.1002/2015GL064052"})
         recognition_file = os.path.join(ocr_dir_test, "reg_recognition_res.txt")
         catalog_file = os.path.join(
             ocr_dir_test, "1010022015GL064052_bibheliotech_V1.txt"
@@ -30,7 +29,7 @@ class TestEntitiesFinder:
         assert os.path.isfile(catalog_file)
 
     def test_catalog_not_empty(self, ocr_dir_test):
-        entities_finder(ocr_dir_test, DOI="10.1002/2015GL064052")
+        entities_finder(ocr_dir_test, doc_meta_info={"doi": "10.1002/2015GL064052"})
         catalog_file = os.path.join(
             ocr_dir_test, "1010022015GL064052_bibheliotech_V1.txt"
         )
@@ -39,7 +38,7 @@ class TestEntitiesFinder:
             assert len(_r_content) == 52
 
     def test_no_image_as_syn(self, ocr_dir_test):
-        entities_finder(ocr_dir_test, DOI="10.1002/2015GL064052")
+        entities_finder(ocr_dir_test, doc_meta_info={"doi": "10.1002/2015GL064052"})
         catalog_file = os.path.join(
             ocr_dir_test, "1010022015GL064052_bibheliotech_V1.txt"
         )
@@ -55,7 +54,7 @@ class TestEntitiesFinder:
     def test_inst_recognition(self, article_as_str, data_frames):
         inst_dict = data_frames[DataBankSheet.INSTR]
         inst_dict_list = inst_recognition(article_as_str, inst_dict)
-        assert len(inst_dict_list) == 56
+        assert len(inst_dict_list) == 12
 
     def test_clean_sats(self, article_as_str, data_frames):
         sat_dict = data_frames[DataBankSheet.SATS]
@@ -75,7 +74,7 @@ class TestEntitiesFinder:
         new_sat_dict_list = clean_sats_inside_insts(sat_dict_list, inst_dict_list)
         _final_links = make_final_links(new_sat_dict_list, inst_list, article_as_str)
         assert len(_final_links) == 39
-        assert list(_final_links[0][0].keys()) == ["end", "start", "text", "type"]
+        assert list(_final_links[0][0].keys()) == ["start", "end", "text", "type"]
         for fl in _final_links:
             assert len(fl) == 2
 
@@ -111,6 +110,9 @@ class TestEntitiesFinder:
         tmp, sat_occ = add_sat_occurrence(final_with_syns, sutime_json)
         published_date = "2015-12-01T00:00:00Z"
         tmp, sat_closest = closest_duration(tmp, sat_occ, data_frames, published_date)
+        for _dict in sat_closest:
+            assert _dict[0]["type"] == "sat"
+            assert _dict[2]["type"] == "DURATION"
         assert len(sat_closest) == 39
         assert len(tmp) == 45