+++ tools/__init__.py
@@ -2,14 +2,9 @@ import glob
 import json
 import copy
 import os
+import re
 
-
-class ToolsError(Exception):
-    """Tools exception"""
-
-    def __init__(self, message="Tools Error"):
-        self.message = message
-        super().__init__(self.message)
+from tools.tools_errors import ToolsValueError, ToolsFileError
 
 
 class RawDumper:
@@ -25,9 +20,9 @@ class RawDumper:
     def dump_to_raw(self, struct_to_dump, message, folder):
         # append message to struct for later reading
         entitled_struct = copy.deepcopy(struct_to_dump)
-        entitled_struct.append(message)
+        entitled_struct.append(f"{self.dump_step}- {message}")
         with open(
-            folder + "/" + f"raw{self.dump_step}_{self.name}.json", "w"
+            os.path.join(folder, f"raw{self.dump_step}_{self.name}.json"), "w"
         ) as raw_file:
             raw_file.write(json.dumps(entitled_struct, sort_keys=True, indent=4))
         self.dump_step = self.dump_step + 1
@@ -35,7 +30,9 @@ class RawDumper:
 
 # TODO: shall we move this to models.paper ?
 class StepLighter:
+
     def __init__(self, ocr_dir, step_num=0, enlight_mode="sutime"):
+        self._all_captions = []
         self.step = int(step_num)
         self.ocr_dir = ocr_dir
         self.enlight_mode = enlight_mode
@@ -61,9 +58,22 @@ class StepLighter:
 
         @return: number of steps in the directory
         """
+        if self._all_captions:
+            return self._all_captions
         jsonfiles_pattern = os.path.join(self.ocr_dir, f"raw*_{self.enlight_mode}.json")
-        all_files = glob.glob(jsonfiles_pattern, recursive=True)
-        return len(all_files)
+        all_files = glob.glob(jsonfiles_pattern)
+        # sort list by step number extracted from file name
+        all_files.sort(
+            key=lambda path: int(re.sub(r".*raw(\d+)_.*\.json", r"\1", path))
+        )
+        # now build the list of captions
+        for f in all_files:
+            with open(f) as json_fd:
+                json_content = json.load(json_fd)
+                step_caption = json_content.pop()
+                self._all_captions.append(step_caption)
+
+        return self._all_captions
 
     @property
     def json_filepath(self):
@@ -74,7 +84,7 @@ class StepLighter:
         print(jsonfile_pattern)
         _json_filepath = glob.glob(jsonfile_pattern, recursive=True)[0]
         if not os.path.isfile(_json_filepath):
-            raise ToolsError(f"No such file {_json_filepath}")
+            raise ToolsFileError(f"No such file {_json_filepath}")
         return _json_filepath
 
     def enlight_step(self):
@@ -93,10 +103,6 @@ class StepLighter:
 
         self.caption_content = step_caption
 
-        for i, elemnt in enumerate(json_content):
-            if type(elemnt) == list:
-                json_content[i] = elemnt[0]
-
         self.enlighted_txt_content = enlight_txt(txt_content, json_content)
 
 
@@ -110,23 +116,60 @@ def enlight_txt(txt_content, json_content):
 
     # json content should be a list of dicts.
     # if not, it is a list of lists of dicts.
-    # in this second case, create a list of the first occurrence of second level
+    # in any case, create a list of any dict containing the 'type' key
 
     # flatten dict list
-    for i, elemnt in enumerate(json_content):
-        if type(elemnt) == list:
-            json_content[i] = elemnt[0]
+    flattened_content = []
+    for elemnt in json_content:
+        if type(elemnt) is dict:
+            flattened_content.append(elemnt)
+        elif type(elemnt) is list:
+            for nested_elmnt in elemnt:
+                flattened_content.append(nested_elmnt)
+        else:
+            raise ToolsValueError(
+                f"No such type allowed in json struct: {type(elemnt)}"
+            )
+    # filter only dict with type key
+    filtered_content = [
+        elmnt
+        for elmnt in flattened_content
+        if type(elmnt) is dict and "type" in elmnt.keys()
+    ]
 
+    filtered_content.sort(key=lambda x: x["start"])
+
+    # remove duplicates
+    uniq_content = []
+    for elmnt in filtered_content:
+        if elmnt not in uniq_content:
+            uniq_content.append(elmnt)
     res_txt = txt_content[:]
     running_offset = 0
-    for i, sutime_struct in enumerate(json_content):
-        if type(sutime_struct) == dict and "type" not in sutime_struct.keys():
-            sutime_struct["type"] = "notype"
+    for i, content_struct in enumerate(uniq_content):
+        content_type = content_struct["type"]
+        content_title = f'Text: {content_struct["text"]}'
+        if content_type == "DURATION":
+            if "value" in content_struct.keys():
+                if "begin" in content_struct["value"]:
+                    content_title += f'&#10;Begin: {content_struct["value"]["begin"]}'
+                if "end" in content_struct["value"]:
+                    content_title += f'&#10;End: {content_struct["value"]["end"]}'
+        elif content_type == "sat":
+            if "D" in content_struct.keys():
+                content_title += f'&#10;D: {content_struct["D"]}'
+            if "R" in content_struct.keys():
+                content_title += f'&#10;R: {content_struct["R"]}'
+            if "SO" in content_struct.keys():
+                content_title += f'&#10;SO: {content_struct["SO"]}'
+            if "conf" in content_struct.keys():
+                content_title += f'&#10;conf: {content_struct["conf"]}'
+            # pass
 
-        opening_tag = f'<span class="highlight {sutime_struct["type"]}" title="{sutime_struct["text"]}">'
+        opening_tag = f'<span class="highlight {content_type}" title="{content_title}">'
         closing_tag = "</span>"
-        start = int(sutime_struct["start"] + running_offset)
-        end = int(sutime_struct["end"] + running_offset)
+        start = int(content_struct["start"]) + running_offset
+        end = int(content_struct["end"]) + running_offset
         opener = res_txt[0:start]
         inner = opening_tag[:] + res_txt[start:end] + closing_tag[:]
         closer = res_txt[end:-1]